+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Enhancements
 
+- [Semantic Field] Support the sparse two phase processor for the semantic field.
+
 ### Bug Fixes
 
 ### Infrastructure